Feature: Implement full extractor design
The extractor inspect labels provided in the input folder (in_folder) to
create the required folders hierarchy in the output folder (out_folder).
It will then iterate over all executables and use the feature extractors
listed in the conf_file to extract the features. It will finally check
the type of the feature extractor to know how to save it on disk.